Components were fixed
