+++ src/components/Filter/Filter.jsx
@@ -1,7 +1,16 @@
 import React from 'react';
+import { useSelector, useDispatch } from 'react-redux';
+import { setFilter } from 'redux/filterSlice';
 import PropTypes from 'prop-types';
 
-const Filter = ({ value, onChange }) => {
+const Filter = () => {
+  const dispatch = useDispatch();
+  const filter = useSelector(state => state.filter);
+
+  const handleFilter = ({ target: { value } }) => {
+    dispatch(setFilter(value));
+  };
+
   return (
     <div className="formWrapper">
       <label>
@@ -10,8 +19,8 @@ const Filter = ({ value, onChange }) => {
           type="text"
           name="filter"
           className="form-control"
-          onChange={onChange}
-          value={value}
+          onChange={handleFilter}
+          value={filter}
         />
       </label>
     </div>
